refactor(dom): extract DomIndexer from DomModel
